[BE] - FEAT : 결식 아동 확인 API (S07P22D101-60)

+++ birdmeal/src/main/java/com/backend/birdmeal/entity/StarvingChildEntity.java
@@ -25,6 +25,6 @@ public class StarvingChildEntity {
 
     @Basic
     @Column(name="child_card_num")
-    private int childCardNum;
+    private Long childCardNum;
 
 }
